+++ src/views/List.vue
@@ -3,19 +3,23 @@ import { ref } from 'vue'
 const listData = [
   {
     title: 'Vue Amazing UI Title 1',
-    description: 'Vue Amazing UI, An Amazing Vue3 UI Components Library.'
+    description: 'Vue Amazing UI, An Amazing Vue3 UI Components Library.',
+    content: 'content 1'
   },
   {
     title: 'Vue Amazing UI Title 2',
-    description: 'Vue Amazing UI, An Amazing Vue3 UI Components Library.'
+    description: 'Vue Amazing UI, An Amazing Vue3 UI Components Library.',
+    content: 'content 2'
   },
   {
     title: 'Vue Amazing UI Title 3',
-    description: 'Vue Amazing UI, An Amazing Vue3 UI Components Library.'
+    description: 'Vue Amazing UI, An Amazing Vue3 UI Components Library.',
+    content: 'content 3'
   },
   {
     title: 'Vue Amazing UI Title 4',
-    description: 'Vue Amazing UI, An Amazing Vue3 UI Components Library.'
+    description: 'Vue Amazing UI, An Amazing Vue3 UI Components Library.',
+    content: 'content 4'
   }
 ]
 const simpleList = [
@@ -40,6 +44,14 @@ const sizeOptions = [
   }
 ]
 const size = ref('middle')
+
+const count = 3
+const fakeDataUrl = `https://randomuser.me/api/?results=${count}&inc=name,gender,email,nat,picture&noinfo`
+
+const initLoading = ref(true)
+const loading = ref(false)
+const data = ref([])
+const list = ref([])
 </script>
 <template>
   <div>
@@ -55,20 +67,6 @@ const size = ref('middle')
         </template>
       </ListItem>
     </List>
-    <a-list item-layout="horizontal" :data-source="listData">
-      <template #renderItem="{ item }">
-        <a-list-item>
-          {{ item.title }}
-          <a-list-item-meta
-            description="Ant Design, a design language for background applications, is refined by Ant UED Team"
-          >
-            <template #avatar>
-              <a-avatar src="https://joeschmoe.io/api/v1/random" />
-            </template>
-          </a-list-item-meta>
-        </a-list-item>
-      </template>
-    </a-list>
     <h2 class="mt30 mb10">带边框列表</h2>
     <List bordered>
       <ListItem v-for="(data, index) in listData" :key="index" :title="data.title">
@@ -108,5 +106,104 @@ const size = ref('middle')
         </ListItem>
       </List>
     </Flex>
+    <h2 class="mt30 mb10">列表添加操作项</h2>
+    <List bordered>
+      <ListItem v-for="(data, index) in listData" :key="index" :title="data.title">
+        <template #avatar>
+          <Avatar src="https://cdn.jsdelivr.net/gh/themusecatcher/resources@0.0.5/1.jpg" />
+        </template>
+        <template #description>
+          {{ data.description }}
+        </template>
+        {{ data.content }}
+        <template #actions>
+          <a>edit</a>
+          <a>more</a>
+          <a>delete</a>
+        </template>
+      </ListItem>
+    </List>
+    <a-list bordered item-layout="horizontal" :data-source="listData">
+      <template #renderItem="{ item }">
+        <a-list-item>
+          <a-list-item-meta
+            description="Ant Design, a design language for background applications, is refined by Ant UED Team"
+          >
+            <template #title>
+              <a href="https://www.antdv.com/">{{ item.title }}</a>
+            </template>
+            <template #avatar>
+              <a-avatar src="https://joeschmoe.io/api/v1/random" />
+            </template>
+          </a-list-item-meta>
+          <div>content</div>
+          <template #actions>
+            <a key="list-loadmore-edit">edit</a>
+            <a key="list-loadmore-more">more</a>
+          </template>
+        </a-list-item>
+      </template>
+    </a-list>
+    <h2 class="mt30 mb10">自定义样式</h2>
+    <List>
+      <ListItem
+        :avatar-style="{ alignSelf: 'center' }"
+        :title-style="{ fontSize: '20px' }"
+        :description-style="{ fontSize: '16px' }"
+        :content-style="{ color: '#f50' }"
+        v-for="(data, index) in listData"
+        :key="index"
+        :title="data.title"
+      >
+        <template #avatar>
+          <Avatar size="large" src="https://cdn.jsdelivr.net/gh/themusecatcher/resources@0.0.5/1.jpg" />
+        </template>
+        <template #description>
+          {{ data.description }}
+        </template>
+        {{ data.content }}
+        <template #actions>
+          <a>edit</a>
+          <a>more</a>
+        </template>
+      </ListItem>
+    </List>
+    <!-- <h2 class="mt30 mb10">加载更多</h2>
+    <a-list
+      class="demo-loadmore-list"
+      :loading="initLoading"
+      item-layout="horizontal"
+      :data-source="list"
+    >
+      <template #loadMore>
+        <div
+          v-if="!initLoading && !loading"
+          :style="{ textAlign: 'center', marginTop: '12px', height: '32px', lineHeight: '32px' }"
+        >
+          <a-button @click="onLoadMore">loading more</a-button>
+        </div>
+      </template>
+      <template #renderItem="{ item }">
+        <a-list-item>
+          <template #actions>
+            <a key="list-loadmore-edit">edit</a>
+            <a key="list-loadmore-more">more</a>
+          </template>
+          <a-skeleton avatar :title="false" :loading="!!item.loading" active>
+            <a-list-item-meta
+              description="Ant Design, a design language for background applications, is refined by Ant UED Team"
+            >
+              <template #title>
+                <a href="https://www.antdv.com/">{{ item.name.last }}</a>
+              </template>
+              <template #avatar>
+                <a-avatar :src="item.picture.large" />
+              </template>
+            </a-list-item-meta>
+            <div>content</div>
+          </a-skeleton>
+        </a-list-item>
+      </template>
+    </a-list> -->
   </div>
 </template>